restructure file parsing folder by format and install yaml parser npm package
