+++ package.json
@@ -53,7 +53,8 @@
         "vscode-languageserver": "^9.0.1",
         "vscode-languageserver-textdocument": "^1.0.12",
         "vscode-uri": "^3.1.0",
-        "watcher": "^2.3.1"
+        "watcher": "^2.3.1",
+        "yaml": "^2.9.0"
     },
     "devDependencies": {
         "@types/chai": "^5.2.2",
