+++ shared/fileParsing/external/bruFormat/getSequenceFieldFromMetaBlock.ts
