+++ shared/fileParsing/external/bruFormat/parseSequenceFromMetaBlock.ts
