+++ shared/fileParsing/external/bruFormat/util/getBlockStartPatternByName.ts
