+++ shared/fileParsing/external/bruFormat/util/getContentRangeForArrayOrDictionaryBlock.ts
