+++ shared/fileParsing/external/bruFormat/util/getNonBlockSpecificBlockStartPattern.ts
