+++ shared/fileParsing/internal/bruFormat/parseDictionaryBlock.ts
@@ -10,8 +10,8 @@ import {
     DictionaryBlockDescription,
     DictionaryBlockTypeAnnotation,
     DictionaryBlockTypeAnnotationValue,
-} from "../..";
-import { getContentRangeForArrayOrDictionaryBlock } from "../external/shared/util/getContentRangeForArrayOrDictionaryBlock";
+} from "../../..";
+import { getContentRangeForArrayOrDictionaryBlock } from "../../external/bruFormat/util/getContentRangeForArrayOrDictionaryBlock";
 
 type ParsedLine =
     | DictionaryBlockArrayField
